Threadsafety improvements using thread local storage
The _opener global object defined in the _opener module was wrapped in a
subclass of threading.local to improve mechanize threadsafety. With this
change that object is no longer shared between threads and is not required
to synchronize the execution of urlopen and urlretrieve.

This modified version of mechanize was used successfully as backend for a stress
test based on the multi-mechanize framework (*) in which up to 512 agents per
user group were active.

(*) http://code.google.com/p/multi-mechanize/

Signed-off-by: Dhionel Díaz <ddiaz@cenditel.gob.ve>